pyproject.toml
[project]
name = "hw2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-search-results>=2.4.2",
    "langchain>=1.2.11",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.18",
    "langchain-google-genai>=4.2.1",
    "langgraph>=1.1.0",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
]
